Split files into gui and database portion
